refactoring for schedule interface
