+++ src/Marcoshoya/MarquejogoBundle/Component/Schedule/ISchedule.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Marcoshoya\MarquejogoBundle\Component\Schedule;
+
+/**
+ * Schedule Interface
+ * 
+ * @author Marcos Lazarin <marcoshoya at gmail dot com>
+ */
+interface ISchedule
+{
+    //put your code here
+}
