+++ src/Marcoshoya/MarquejogoBundle/Resources/config/services.yml
@@ -24,8 +24,8 @@ services:
         class: Marcoshoya\MarquejogoBundle\Service\SearchService
         parent: marcoshoya_marquejogo.service.base
 
-    marcoshoya_marquejogo.service.product:
-        class: Marcoshoya\MarquejogoBundle\Service\ProductService
+    marcoshoya_marquejogo.service.schedule:
+        class: Marcoshoya\MarquejogoBundle\Service\ScheduleService
         parent: marcoshoya_marquejogo.service.base
 
     marcoshoya_marquejogo.service.book:
